charter: Add a section on OCI Projects
Because "everything the OCI is working on" is usually more important
than distinguishing between specifications and implementations.  And
"OCI Specifications and/or OCI Projects" (which was used several
times) didn't list the certification tooling or application described
in the Trademark Board's old f.i ("creating the OCI trademarks
associated with OCI Specifications, OCI Projects, the Open Container
Format (OCF) or OCI Certified Solution").  It's better to just lump
all these efforts together under "OCI Projects" and then reference
them as a group later in the charter.