charter: Make per-project TDCs explicit
If the set of spec maintainers and the set of runC maintainers is
identical, spec maintainers who are also runC developers might be
tempted to skew the spec towards already-implemented-in-runC
semantics, despite the existence of potentially cleaner models in
other implementations.  To avoid bias in the spec, it makes sense to
explicitly declare per-project TDCs.  That allows developers
associated with any implemention (whether or not that implementation
is an OCI Project) to be on the spec TDC without requiring them to
also assume responsibility for runC itself.

For background discussion, see [1].

[1]: https://groups.google.com/a/opencontainers.org/forum/#!topic/dev/Tjq5QJ_eez0
  In case that link dies, or folks want to look up the conversation in
  their local mail archives, here's some information from the message
  spawning that thread:
  Date: Mon, 3 Aug 2015 16:20:04 +0800
  Message-ID: <CAHKD89TqBQvbcdvwt-g3J--sDyq=P50qi6CytokNJznVsei7YQ@mail.gmail.com>
  Subject: Contribute to OCI
  From: Thibault Bronchain
  To: dev@opencontainers.org