charter: Allow the TOB to create/destroy OCI Projects (and their TDCs)
The previous commit added per-project TDCs.  With per-project
institutions, it makes sense for the OCI to have a mechanism for
adding/removing projects, and the TOB seems like the obvious choice to
do that (the LF board probably doesn't care about this level of
granularity, the other TDCs are focused on other projects, and OCI
Members probably shouldn't be able to launch OCI Projects
independently).