charter: Define "TDC Members" and remove non-example "maintainers"
We never formally define "TDC Maintainers".  They're just an example
role, and 4.b.viii.1 allows the TDC to define it's own roles.  This
commit replaces normative references to "maintainers" and
"contributors" with TDC Members, and tries to clarify TDC Role-holders
as a subset of TDC Members.

It's unclear how TDC Roles are established and maintained (e.g. maybe
only Maintainers get to vote on adding new Roles or inducting new
Maintainers).  Presumably initial the set of TDC Members can reach an
agreement on the initial rule adoption, and further rule changes can
be governed by that agreement.  In the event that they cannot reach an
agreement on the initial rule set, I guess they can appeal to the TOB
(5.a), although that would have to wait until after the TDC and OCI
Members elect the TOB ;).